+++ cuda/src/derivs.cu
@@ -1,6 +1,6 @@
 #include "cuda_defs.cuh"
 
-__device__ void derivs(double x, double y[], double dydx[], float a, float b, float c) {
+__device__ void derivs(double x, double y[], double dydx[], double a, double b, double c) {
     dydx[0] = (-y[1] - y[2]);
     dydx[1] = (y[0] + a * y[1]);
     dydx[2] = (b + y[2] * (y[0] - c));
